+++ automodel/models/auto_model_for_image_text_to_text.py
