+++ automodel/training/init_utils.py
@@ -18,8 +18,9 @@ import signal
 
 import torch
 import torch.distributed
+from dataclasses import dataclass
 
-from nemo_lm.automodel.utils.dist_utils import (
+from automodel.utils.dist_utils import (
     get_local_rank_preinit,
     get_rank_safe,
     get_world_size_safe
