+++ recipes/llama_3_2_1b_hellaswag.yaml
@@ -1,5 +1,5 @@
 training:
-  _target_: nemo_lm.config.common.TrainingConfig
+  _target_: config.common.TrainingConfig
   train_iters: 250
   eval_interval: 1000
   eval_iters: 4
@@ -10,7 +10,7 @@ distributed:
   timeout_minutes: 1
 
 rng:
-  _target_: nemo_lm.automodel.training.rng.StatefulRNG
+  _target_: automodel.training.rng.StatefulRNG
   seed: 1111
   ranked: true
 
@@ -18,10 +18,10 @@ model:
   _target_: transformers.AutoModelForCausalLM.from_pretrained
   pretrained_model_name_or_path: meta-llama/Llama-3.2-1B
 
-loss_fn: nemo_lm.automodel.loss.masked_ce.masked_cross_entropy
+loss_fn: automodel.loss.masked_ce.masked_cross_entropy
 
 dataset:
-  _target_: nemo_lm.automodel.datasets.hellaswag.HellaSwag
+  _target_: automodel.datasets.hellaswag.HellaSwag
   path_or_dataset: rowan/hellaswag
   split: train
   tokenizer:
@@ -30,11 +30,11 @@ dataset:
 
 dataloader:
   _target_: torchdata.stateful_dataloader.StatefulDataLoader
-  collate_fn: nemo_lm.automodel.datasets.utils.default_collater
+  collate_fn: automodel.datasets.utils.default_collater
   batch_size: 1
 
 validation_dataset:
-  _target_: nemo_lm.automodel.datasets.hellaswag.hellaswag
+  _target_: automodel.datasets.hellaswag.hellaswag
   path_or_dataset: rowan/hellaswag
   split: train
   tokenizer:
@@ -53,7 +53,7 @@ optimizer:
   # min_lr: 1.0e-5
 
 scheduler:
-  _target_: nemo_lm.automodel.config.SchedulerConfig
+  _target_: automodel.config.SchedulerConfig
   start_weight_decay: 0
   end_weight_decay: 0
   weight_decay_incr_style: constant
@@ -64,7 +64,7 @@ scheduler:
   override_opt_param_scheduler: true
 
 logger:
-  _target_: nemo_lm.config.common.LoggerConfig
+  _target_: config.common.LoggerConfig
   wandb_project: nemo_automodel_sft_loop
   wandb_entity: nvidia
   wandb_exp_name: nemolm_automodel_Rowan_hellaswag_meta-llama_Llama-3.2-1B_gbs_256_seq_len_1024_lr_1.0e-5
@@ -81,7 +81,7 @@ logger:
     - nemo.collections.llm.gpt.data.utils
 
 checkpointer:
-  _target_: nemo_lm.automodel.training.checkpoint.TorchCheckpointer
+  _target_: automodel.training.checkpoint.TorchCheckpointer
   # save_interval: 10000
   # save: /tmp/nemo_run/checkpoints/automodel/Rowan_hellaswag_meta-llama_Llama-3.2-1B_gbs_256_seq_len_1024
   # load: /tmp/nemo_run/checkpoints/automodel/Rowan_hellaswag_meta-llama_Llama-3.2-1B_gbs_256_seq_len_1024
